2022/Lake/Immutable/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bcrypt==4.0.1
bitarray==2.6.0
capstone==5.0.0rc2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colored-traceback==0.3.0
cryptography==39.0.0
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.4
importlib-resources==5.10.0
intervaltree==3.1.0
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
Mako==1.2.4
MarkupSafe==2.1.2
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==23.0
paramiko==3.0.0
parsimonious==0.8.1
pkgutil_resolve_name==1.3.10
plumbum==1.8.1
protobuf==3.19.5
psutil==5.9.4
pwntools==4.9.0
py-solc-x==1.1.1
pycparser==2.21
pycryptodome==3.15.0
pyelftools==0.29
Pygments==2.14.0
PyNaCl==1.5.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.28.1
rlp==2.0.1
ROPGadget==7.2
rpyc==5.3.0
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
toolz==0.12.0
unicorn==2.0.1.post1
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
yarl==1.8.1
zipp==3.9.0
